+++ src/img.c
@@ -1,5 +1,4 @@
 #include "cub3D.h"
-#include <stdio.h>
 
 int	set_img(t_game *game)
 {
